Fix counter/checksum build error
When there were signals named CHECKSUM or COUNTER in build did not finish because it gave STEERING_COMMAND: COUNTER has wrong endianness
and STEERING_STATUS: CHECKSUM starts at wrong bit; errors. It seems that comma has an assert if signals that are previously named does not follow certain logik, so I just change the names, not sure if it is optimal solution.